refactored by eliminating an extra component layer I didn't need :). also added embedded maps using destination lat and lng to html export.
